+++ src/main/java/de/caritas/cob/userservice/api/adapters/web/dto/UserRegistrationDTO.java
@@ -14,13 +14,13 @@ public interface UserRegistrationDTO {
 
   String getConsultantId();
 
-  Integer getMainTopicId();
+  Long getMainTopicId();
 
   String getUserGender();
 
   Integer getUserAge();
 
-  Collection<Integer> getTopicIds();
+  Collection<Long> getTopicIds();
 
   String getCounsellingRelation();
 }
